Improve dataclass example.
